Strangler extraction #12: emulator ops -> zxnu_emulator_ops.py
Move the emulator + self-update operation layer (~1.5k lines, pure
closure definitions with no construction-time side effects) out of
MainWindow.__init__ into build_emulator_ops(host, ...): the CSpect/MAME
settings-combo setters, open_cspect_configuration_file, launch_cspect,
launch_mame, the MAME install/update chain, the zx-next-unite
self-update chain, the .sync dotN version advisory, the CSpect update
chain and _wire_viewer_emulators.

Same builder-function seam: self -> host via a tokenize NAME pass;
configuration/save/get_tuple_value/button-state/_update_mame_controls
injected by identity, _start_load_image_hint_animation /
add_main_log_window / execute_shell_command as forwarding lambdas (bound
later in __init__), right_disk_image_explorer_content via the
_right_disk_content getter hook. The 16 closures the widget wiring
consumes by bare name are exposed as host attributes and re-bound to
bare locals at the call site. zxnu_main.py 10,925 -> 9,467 lines.

Note: tests/test_classic_sync.py flakes ~1-in-3 under back-to-back runs
with a WinError 10054 shutdown race — pre-existing (reproduced without
these changes, zxnu_workers.py untouched), tracked separately.
