+++ tests/test_pane_imports.py
@@ -134,6 +134,9 @@ PANES = [
     ("zxnu_favorites_pane",
      ["build_favorites_helpers", "build_favorites_pane",
       "build_favorites_ops"], []),
+    # Extraction #12: the emulator + self-update operation layer (CSpect/MAME
+    # setters + launchers, MAME/CSpect/app update chains, viewer wiring).
+    ("zxnu_emulator_ops", "build_emulator_ops", []),
 ]
 
 for modname, funcnames, _outputs in PANES:
